Minor refactor: Renamed menu-package to GUI, as this might be more fitting
