+++ src/main/java/inf112/skeleton/app/gameStates/MainMenu/State_MainMenu.java
@@ -4,18 +4,14 @@ package inf112.skeleton.app.gameStates.MainMenu;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.google.gson.Gson;
-import inf112.skeleton.app.ChatGUI;
 import inf112.skeleton.app.gameStates.GameState;
 import inf112.skeleton.app.gameStates.GameStateManager;
 import inf112.skeleton.app.gameStates.Playing.State_Playing;
 import inf112.skeleton.app.Action.Action;
-import inf112.skeleton.app.menu.Menu;
+import inf112.skeleton.app.GUI.Menu;
 import inf112.skeleton.common.packet.LoginPacket;
 import inf112.skeleton.common.packet.Packet;
 import io.netty.channel.Channel;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class State_MainMenu extends GameState {
     private Menu menu;
